Debug and add tests

+++ src/instruction/srl.cpp
@@ -8,7 +8,8 @@ Simulator::State Simulator::srl(Instruction inst)
     auto op = decodeR(inst);
 
     new_state.pc += 4;
-    new_state.reg.at(op.rd) = m_state_iter->reg.at(op.rs) >> op.shamt;
+    new_state.reg.at(op.rd) = static_cast<int32_t>(
+        static_cast<uint32_t>(m_state_iter->reg.at(op.rs)) >> op.shamt);
 
     return new_state;
 }
